Migrate endpoints from server to api
