+++ lib/api/shared/src/main/scala/net/wiringbits/api/endpoints/EnvironmentConfigEndpoints.scala
@@ -1,0 +1,27 @@
+package net.wiringbits.api.endpoints
+
+import net.wiringbits.api.models
+import net.wiringbits.api.models.GetEnvironmentConfig
+import sttp.tapir.*
+import sttp.tapir.json.play.*
+
+object EnvironmentConfigEndpoints {
+  private val baseEndpoint = endpoint
+    .in("environment-config")
+    .tag("Misc")
+    .errorOut(errorResponseErrorOut)
+
+  val getEnvironmentConfig: Endpoint[Unit, Unit, models.ErrorResponse, GetEnvironmentConfig.Response, Any] =
+    baseEndpoint.get
+      .out(
+        jsonBody[GetEnvironmentConfig.Response]
+          .description("Got the config values")
+          .example(GetEnvironmentConfig.Response("siteKey"))
+      )
+      .summary("Get the config values for the current environment")
+      .description("These values are required by the frontend app to interact with the backend")
+
+  val routes: List[Endpoint[_, _, _, _, _]] = List(
+    getEnvironmentConfig
+  )
+}
