+++ server/src/main/scala/controllers/AuthController.scala
@@ -1,6 +1,7 @@
 package controllers
 
 import net.wiringbits.actions.*
+import net.wiringbits.api.endpoints.AuthEndpoints
 import net.wiringbits.api.models.*
 import net.wiringbits.common.models.*
 import org.slf4j.LoggerFactory
@@ -74,78 +75,9 @@ class AuthController @Inject() (
 
   def routes: List[ServerEndpoint[AkkaStreams with WebSockets, Future]] = {
     List(
-      AuthController.login.serverLogic(login),
-      AuthController.getCurrentUser.serverLogic(me),
-      AuthController.logout.serverLogic(logout)
+      AuthEndpoints.login.serverLogic(login),
+      AuthEndpoints.getCurrentUser.serverLogic(me),
+      AuthEndpoints.logout.serverLogic(logout)
     )
   }
-}
-
-object AuthController {
-  import sttp.tapir.*
-  import sttp.tapir.json.play.*
-
-  private val baseEndpoint = endpoint
-    .in("auth")
-    .tag("Auth")
-    .errorOut(errorResponseErrorOut)
-
-  private val login = baseEndpoint.post
-    .in("login")
-    .in(
-      jsonBody[Login.Request].example(
-        Login.Request(
-          email = Email.trusted("alexis@wiringbits.net"),
-          password = Password.trusted("notSoWeakPassword"),
-          captcha = Captcha.trusted("captcha")
-        )
-      )
-    )
-    .out(
-      jsonBody[Login.Response]
-        .description("Successful login")
-        .example(
-          Login.Response(
-            id = UUID.fromString("3fa85f64-5717-4562-b3fc-2c963f66afa6"),
-            name = Name.trusted("Alexis"),
-            email = Email.trusted("alexis@wiringbits.net")
-          )
-        )
-    )
-    .out(setUserIdCookie)
-    .errorOut(oneOf(HttpErrors.badRequest))
-    .summary("Log into the app")
-    .description("Sets a session cookie to authenticate the following requests")
-
-  private val logout = baseEndpoint.post
-    .in("logout")
-    .in(userIdCookie)
-    .out(jsonBody[Logout.Response].description("Successful logout").example(Logout.Response()))
-    .out(setUserIdCookie)
-    .errorOut(oneOf(HttpErrors.badRequest))
-    .summary("Logout from the app")
-    .description("Clears the session cookie that's stored securely")
-
-  private val getCurrentUser = baseEndpoint.get
-    .in("me")
-    .in(userIdCookie)
-    .out(
-      jsonBody[GetCurrentUser.Response]
-        .description("Got user details")
-        .example(
-          GetCurrentUser.Response(
-            id = UUID.fromString("3fa85f64-5717-4562-b3fc-2c963f66afa6"),
-            name = Name.trusted("Alexis"),
-            email = Email.trusted("alexis@wiringbits.net"),
-            createdAt = Instant.parse("2021-01-01T00:00:00Z")
-          )
-        )
-    )
-    .summary("Get the details for the authenticated user")
-
-  val routes: List[Endpoint[_, _, _, _, _]] = List(
-    login,
-    logout,
-    getCurrentUser
-  )
 }
