+++ server/src/main/scala/controllers/EnvironmentConfigController.scala
@@ -1,6 +1,7 @@
 package controllers
 
 import net.wiringbits.actions.GetEnvironmentConfigAction
+import net.wiringbits.api.endpoints.EnvironmentConfigEndpoints
 import net.wiringbits.api.models.{ErrorResponse, GetEnvironmentConfig}
 import org.slf4j.LoggerFactory
 import sttp.capabilities.WebSockets
@@ -24,29 +25,6 @@ class EnvironmentConfigController @Inject() (
   }
 
   def routes: List[ServerEndpoint[AkkaStreams with WebSockets, Future]] = {
-    List(EnvironmentConfigController.getEnvironmentConfig.serverLogic(_ => getEnvironmentConfig))
+    List(EnvironmentConfigEndpoints.getEnvironmentConfig.serverLogic(_ => getEnvironmentConfig))
   }
-}
-
-object EnvironmentConfigController {
-  import sttp.tapir.*
-  import sttp.tapir.json.play.*
-
-  private val baseEndpoint = endpoint
-    .in("environment-config")
-    .tag("Misc")
-    .errorOut(errorResponseErrorOut)
-
-  private val getEnvironmentConfig = baseEndpoint.get
-    .out(
-      jsonBody[GetEnvironmentConfig.Response]
-        .description("Got the config values")
-        .example(GetEnvironmentConfig.Response("siteKey"))
-    )
-    .summary("Get the config values for the current environment")
-    .description("These values are required by the frontend app to interact with the backend")
-
-  val routes: List[Endpoint[_, _, _, _, _]] = List(
-    getEnvironmentConfig
-  )
 }
