+++ server/src/main/scala/controllers/package.scala
@@ -37,37 +37,4 @@ package object controllers {
       logger.debug(s"Error response while handling a request: ${ex.getMessage}", ex)
       Left(ErrorResponse(ex.getMessage))
   }
-
-  // TODO: better name?
-  object HttpErrors {
-    val badRequest: EndpointOutput.OneOfVariant[Unit] = oneOfVariant(
-      statusCode(StatusCode.BadRequest).description("Invalid or missing arguments")
-    )
-
-    val unauthorized: EndpointOutput.OneOfVariant[Unit] = oneOfVariant(
-      statusCode(StatusCode.Unauthorized).description("Invalid or missing authentication")
-    )
-  }
-
-  val adminCookie: EndpointInput.Cookie[String] = cookie[String]("X-Forwarded-User")
-    .default("Unknown")
-    .schema(_.hidden(true))
-
-  val adminAuth: EndpointInput.Auth[String, AuthType.Http] = auth
-    .basic[String]()
-    .securitySchemeName("Basic authorization")
-    .description("Admin credentials")
-
-  val userIdCookie: EndpointInput.Cookie[Option[UUID]] = cookie[Option[UUID]]("userId")
-    .description("User session")
-    .schema(_.hidden(true))
-
-  val setUserIdCookie: EndpointIO.Header[CookieValueWithMeta] = setCookie("userId")
-    .description("User session")
-    .schema(_.hidden(true))
-
-  val errorResponseErrorOut: EndpointIO.Body[String, ErrorResponse] = jsonBody[ErrorResponse]
-    .description("Error response")
-    .example(ErrorResponse("Unauthorized: Invalid or missing authentication"))
-    .schema(_.hidden(true))
 }
